Move generateSKID from issuance to ca
